TeamSelection component working as expected. Testing still pending for persisting the changes.

+++ TeamWeeklyStatus.WebApi/Controllers/WeeklyStatusController.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using TeamWeeklyStatus.Domain.DTOs;
 using TeamWeeklyStatus.Application.Interfaces;
 using TeamWeeklyStatus.WebApi.DTOs;
 
@@ -19,7 +18,7 @@ namespace TeamWeeklyStatus.WebApi.Controllers
         [HttpPost("GetByMemberIdAndStartDate", Name = "GetWeeklyStatusByMemberByStartDate")]
         public async Task<IActionResult> GetWeeklyStatusByMemberByStartDate([FromBody] WeeklyStatusGetRequest request)
         {
-            var weeklyStatus = await _weeklyStatusService.GetWeeklyStatusByMemberByStartDateAsync((int)request.MemberId, request.TeamId, request.WeekStartDate);
+            var weeklyStatus = await _weeklyStatusService.GetWeeklyStatusByMemberByStartDateAsync((int)request.MemberId, (int)request.TeamId, request.WeekStartDate);
             if (weeklyStatus == null)
             {
                 return NotFound();
